Create Controller

+++ src/main/java/com/khrapkov/utask/dto/request/RequestPaymentDto.java
@@ -2,9 +2,12 @@ package com.khrapkov.utask.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.util.Currency;
+
 @Data
 public class RequestPaymentDto {
     private String sender;
     private String recipient;
-    private Double amount;
+    private long amount;
 }
